refactor(core): flatten packages
