+++ client/src/main/java/net/lapidist/colony/client/Colony.java
@@ -11,7 +11,7 @@ import net.lapidist.colony.client.network.GameClient;
 import net.lapidist.colony.server.GameServer;
 import net.lapidist.colony.server.GameServerConfig;
 import net.lapidist.colony.config.ColonyConfig;
-import net.lapidist.colony.core.events.Events;
+import net.lapidist.colony.events.Events;
 import net.lapidist.colony.client.events.GameInitEvent;
 
 import java.io.IOException;
