+++ client/src/main/java/net/lapidist/colony/client/events/GameInitEvent.java
@@ -1,6 +1,6 @@
 package net.lapidist.colony.client.events;
 
-import net.lapidist.colony.core.events.SimpleEvent;
+import net.lapidist.colony.events.SimpleEvent;
 
 /**
  * Event fired when the game has finished initialisation.
