+++ client/src/main/java/net/lapidist/colony/client/events/PauseEvent.java
@@ -1,6 +1,6 @@
 package net.lapidist.colony.client.events;
 
-import net.lapidist.colony.core.events.SimpleEvent;
+import net.lapidist.colony.events.SimpleEvent;
 public final class PauseEvent extends SimpleEvent {
 
 }
