--- core/src/main/java/net/lapidist/colony/core/events/package-info.java
@@ -1,1 +1,0 @@
-package net.lapidist.colony.core.events;
