+++ core/src/main/java/net/lapidist/colony/events/Events.java
@@ -1,4 +1,4 @@
-package net.lapidist.colony.core.events;
+package net.lapidist.colony.events;
 
 import net.mostlyoriginal.api.event.common.Event;
 import net.mostlyoriginal.api.event.common.EventSystem;
