+++ core/src/main/java/net/lapidist/colony/events/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Global event utilities.
+ */
+package net.lapidist.colony.events;
