+++ server/src/main/java/net/lapidist/colony/server/GameServer.java
@@ -3,8 +3,8 @@ package net.lapidist.colony.server;
 import com.esotericsoftware.kryonet.Server;
 import net.lapidist.colony.components.state.MapState;
 import net.lapidist.colony.config.ColonyConfig;
-import net.lapidist.colony.core.events.Events;
-import net.lapidist.colony.core.serialization.KryoRegistry;
+import net.lapidist.colony.events.Events;
+import net.lapidist.colony.serialization.KryoRegistry;
 import net.lapidist.colony.network.AbstractMessageEndpoint;
 import net.lapidist.colony.network.MessageHandler;
 import net.lapidist.colony.io.Paths;
