+++ server/src/main/java/net/lapidist/colony/server/commands/TileSelectionCommandHandler.java
@@ -4,7 +4,7 @@ import net.lapidist.colony.components.state.MapState;
 import net.lapidist.colony.components.state.TileData;
 import net.lapidist.colony.components.state.TilePos;
 import net.lapidist.colony.components.state.TileSelectionData;
-import net.lapidist.colony.core.events.Events;
+import net.lapidist.colony.events.Events;
 import net.lapidist.colony.server.events.TileSelectionEvent;
 import net.lapidist.colony.server.services.NetworkService;
