+++ tests/src/test/java/net/lapidist/colony/tests/core/events/EventsTest.java
@@ -1,7 +1,7 @@
 package net.lapidist.colony.tests.core.events;
 
 import net.lapidist.colony.client.events.*;
-import net.lapidist.colony.core.events.Events;
+import net.lapidist.colony.events.Events;
 import net.lapidist.colony.tests.GdxTestRunner;
 import net.mostlyoriginal.api.event.common.EventSystem;
 import net.mostlyoriginal.api.event.common.Subscribe;
