+++ tests/src/test/java/net/lapidist/colony/tests/server/GameServerChatCommandTest.java
@@ -6,7 +6,7 @@ import net.lapidist.colony.components.state.TilePos;
 import net.lapidist.colony.server.GameServer;
 import net.lapidist.colony.server.GameServerConfig;
 import net.lapidist.colony.server.events.TileSelectionEvent;
-import net.lapidist.colony.core.events.Events;
+import net.lapidist.colony.events.Events;
 import net.mostlyoriginal.api.event.common.Subscribe;
 import org.junit.Test;
