+++ tests/src/test/java/net/lapidist/colony/tests/server/GameStateIOMigrationV5Test.java
@@ -3,7 +3,7 @@ package net.lapidist.colony.tests.server;
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.io.Output;
 import net.lapidist.colony.components.state.MapState;
-import net.lapidist.colony.core.serialization.KryoRegistry;
+import net.lapidist.colony.serialization.KryoRegistry;
 import net.lapidist.colony.save.SaveData;
 import net.lapidist.colony.save.SaveVersion;
 import net.lapidist.colony.serialization.SerializationRegistrar;
