+++ tests/src/test/java/net/lapidist/colony/tests/server/events/ServerEventsTest.java
@@ -1,7 +1,7 @@
 package net.lapidist.colony.tests.server.events;
 
 import net.lapidist.colony.server.events.AutosaveEvent;
-import net.lapidist.colony.core.events.Events;
+import net.lapidist.colony.events.Events;
 import net.lapidist.colony.server.events.ShutdownSaveEvent;
 import net.lapidist.colony.server.events.TileSelectionEvent;
 import net.lapidist.colony.server.events.BuildingPlacedEvent;
